Fix multiple bots
