+++ game/src/main/kotlin/world/gregs/voidps/bot/Bots.kt
@@ -1,0 +1,7 @@
+package world.gregs.voidps.bot
+
+import world.gregs.voidps.engine.entity.character.player.Player
+import world.gregs.voidps.engine.entity.contains
+
+val Player.isBot: Boolean
+    get() = contains("bot")
